Firebase Remote Config.
Fixed bug with firebase remote config not being able to retrieve values. This was due to the app's manifest missing the "internet" permission.

Create a Util for Game, to abstract the firebase remote config pieces for all three board types.

Using the values from firebase remote config to set game timer.

BugFix: fixed the color issue for the bottom navigation.