Clean-up of the GameViewModel.kt.
Fixes to onGameFinishComplete in GameViewModel.kt to better handle the use case where a winner hits the back button. Previously the game board was not resetting correctly.

Updated navigation.xml to set the corret launchSingleTop, popUpTo, and popUpToInclusive attributes on actions.

Updated GameFragment2x2.kt, GameFragment3x3.kt, and GameFragment4x4.kt to handle the back button, and to reset the board correctly after a winner.